move even more stuff to release.py
