+++ fab/fabfile.py
@@ -34,7 +34,7 @@ from distutils.util import strtobool
 from fabric import utils
 from fabric.api import run, roles, execute, task, sudo, env, parallel
 from fabric.colors import blue, red
-from fabric.context_managers import settings, cd
+from fabric.context_managers import cd
 from fabric.contrib import files, console
 from fabric.operations import require
 from const import (
@@ -439,66 +439,13 @@ def unlink_current():
         sudo('unlink {}'.format(env.code_current))
 
 
-@parallel
-@roles(ROLES_ALL_SRC)
-def copy_localsettings():
-    sudo('cp {}/localsettings.py {}/localsettings.py'.format(env.code_current, env.code_root))
-
-
-@parallel
-@roles(ROLES_TOUCHFORMS)
-def copy_tf_localsettings():
-    sudo(
-        'cp {}/submodules/touchforms-src/touchforms/backend/localsettings.py '
-        '{}/submodules/touchforms-src/touchforms/backend/localsettings.py'.format(
-            env.code_current, env.code_root
-        ))
-
-
-@parallel
-@roles(ROLES_TOUCHFORMS)
-def copy_formplayer_properties():
-    with settings(warn_only=True):
-        sudo(
-            'cp {}/submodules/formplayer/config/application.properties '
-            '{}/submodules/formplayer/config'.format(
-                env.code_current, env.code_root
-            ))
-
-
-@parallel
-@roles(ROLES_ALL_SRC)
-def copy_components():
-    if files.exists('{}/bower_components'.format(env.code_current)):
-        sudo('cp -r {}/bower_components {}/bower_components'.format(env.code_current, env.code_root))
-    else:
-        sudo('mkdir {}/bower_components'.format(env.code_root))
-
-
-@parallel
-@roles(ROLES_ALL_SRC)
-def copy_node_modules():
-    if files.exists('{}/node_modules'.format(env.code_current)):
-        sudo('cp -r {}/node_modules {}/node_modules'.format(env.code_current, env.code_root))
-    else:
-        sudo('mkdir {}/node_modules'.format(env.code_root))
-
-
-@parallel
-@roles(ROLES_STATIC)
-def copy_compressed_js_staticfiles():
-    if files.exists('{}/staticfiles/CACHE/js'.format(env.code_current)):
-        sudo('mkdir -p {}/staticfiles/CACHE/js'.format(env.code_root))
-        sudo('cp -r {}/staticfiles/CACHE/js {}/staticfiles/CACHE/js'.format(env.code_current, env.code_root))
-
-
 def copy_release_files():
-    execute(copy_localsettings)
-    execute(copy_tf_localsettings)
-    execute(copy_formplayer_properties)
-    execute(copy_components)
-    execute(copy_node_modules)
-    execute(copy_compressed_js_staticfiles)
+    execute(release.copy_localsettings)
+    execute(release.copy_tf_localsettings)
+    execute(release.copy_formplayer_properties)
+    execute(release.copy_components)
+    execute(release.copy_node_modules)
+    execute(release.copy_compressed_js_staticfiles)
 
 
 @task
@@ -507,12 +454,12 @@ def rollback():
     Rolls back the servers to the previous release if it exists and is same
     across servers.
     """
-    number_of_releases = execute(get_number_of_releases)
+    number_of_releases = execute(release.get_number_of_releases)
     if not all(map(lambda n: n > 1, number_of_releases)):
         print red('Aborting because there are not enough previous releases.')
         exit()
 
-    releases = execute(get_previous_release)
+    releases = execute(release.get_previous_release)
 
     unique_releases = set(releases.values())
     if len(unique_releases) != 1:
@@ -530,7 +477,7 @@ def rollback():
         print blue('Exiting.')
         exit()
 
-    exists = execute(ensure_release_exists, unique_release)
+    exists = execute(release.ensure_release_exists, unique_release)
 
     if all(exists.values()):
         print blue('Updating current and restarting services')
@@ -540,27 +487,6 @@ def rollback():
     else:
         print red('Aborting because not all hosts have release')
         exit()
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def get_number_of_releases():
-    with cd(env.root):
-        return int(sudo("wc -l {} | awk '{{ print $1 }}'".format(RELEASE_RECORD)))
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def ensure_release_exists(release):
-    return files.exists(release)
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def get_previous_release():
-    # Gets second to last line in RELEASES.txt
-    with cd(env.root):
-        return sudo('tail -2 {} | head -n 1'.format(RELEASE_RECORD))
 
 
 @task
